--- src/pages/dashboardRequestToBeAnswered/DashboardRequestToBeAnswered.tsx
@@ -1,568 +1,0 @@
-import {
-  Chip,
-  Grid,
-  Typography,
-  Alert,
-  Stack,
-  Paper,
-  Divider,
-  styled,
-  Button,
-} from '@mui/material';
-import { theme } from '@pagopa/mui-italia';
-import TitleBox from '@pagopa/selfcare-common-frontend/components/TitleBox';
-import { useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import ConfirmPage from '../confirmPage/ConfirmPage';
-import RejectPage from '../rejectedPage/RejectPage';
-
-const CustomAlert = styled(Alert)({
-  '& .MuiAlert-icon': {
-    color: '#FFCB46',
-  },
-});
-
-export default function DashboardRequestToBeAnswered() {
-  const { t } = useTranslation();
-  const [showRejectPage, setShowRejectPage] = useState<boolean>();
-  const [showConfirmPage, setShowConfirmPage] = useState<boolean>();
-
-  // TODO: to remove when api will be ready
-  const toValidateState = true;
-
-  const stateCondtition = (state: any, isChipLabelState: boolean, isBgColorChipState: boolean) => {
-    if (isChipLabelState) {
-      switch (state) {
-        case 'approved':
-          return t('requestToBeAnsweredPage.approvedDataChip');
-        case 'rejected':
-          return t('requestToBeAnsweredPage.rejectedDataChip');
-        default:
-          return t('requestToBeAnsweredPage.validationDataChip');
-      }
-    }
-    if (isBgColorChipState) {
-      switch (state) {
-        case 'approved':
-          return 'success.light';
-        case 'rejected':
-          return 'error.light';
-        default:
-          return 'warning.main';
-      }
-    }
-    return undefined;
-  };
-
-  return showRejectPage ? (
-    <RejectPage />
-  ) : showConfirmPage ? (
-    <ConfirmPage />
-  ) : (
-    <Grid container ml={32} sx={{ width: '920px' }}>
-      <Grid item xs={12}>
-        <Grid container direction="row" justifyContent="space-between" alignItems="center" mt={6}>
-          <Grid item>
-            <Typography variant="h4"> {t('requestToBeAnsweredPage.title')} </Typography>
-          </Grid>
-          <Grid item>
-            <Chip
-              sx={{
-                backgroundColor: stateCondtition('toValidate', false, true),
-                height: '30px',
-              }}
-              label={stateCondtition('toValidate', true, false)}
-            />
-          </Grid>
-        </Grid>
-        {toValidateState && (
-          <Grid item xs={12} width="100%" mt={5}>
-            <CustomAlert
-              severity="warning"
-              sx={{
-                fontSize: 'fontSize',
-                backgroundColor: 'background.paper',
-                height: '80px',
-                alignItems: 'center',
-                color: 'colorTextPrimary',
-                borderLeft: 'solid',
-                borderLeftColor: 'warning.main',
-                borderLeftWidth: '4px',
-                width: '100%',
-              }}
-            >
-              {t('requestToBeAnsweredPage.checkPartyInfoAlert')}
-            </CustomAlert>
-          </Grid>
-        )}
-        <Stack spacing={4} mt={4} sx={{ width: '100%' }}>
-          <Paper elevation={8} sx={{ borderRadius: theme.spacing(2) }}>
-            <Grid container sx={{ marginY: 4, marginX: 4 }}>
-              <Grid item xs={12}>
-                <TitleBox
-                  title={t(
-                    'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.title'
-                  )}
-                  variantTitle={'h6'}
-                  mtTitle={1}
-                  mbTitle={5}
-                />
-                <Divider />
-              </Grid>
-              <Grid item xs={12} mt={4}>
-                <Grid container spacing={1}>
-                  {/* businessName */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.businessName'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'businessNameData'} {/* TODO businessName */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* registeredOffice */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.registeredOffice'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'registeredOfficeData'} {/* TODO registeredOffice */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* zipCode */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.zipCode'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'zipCodeData'} {/* TODO zipCode */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* mailPEC */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.mailPEC'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'mailPECData'} {/* TODO mailPEC */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* taxcode */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.taxCode'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'taxCodeData'} {/* TODO taxCodeData */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* vatNumber */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.vatNumber'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'vatNumberData'} {/* TODO vatNumber */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* isGroupPIVA */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.isGroupPIVA'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'isGroupPIVAData'} {/* TODO isGroupPIVA */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* commercialRegisterNumber */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.commercialRegisterNumber'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'commercialRegisterNumberData'} {/* TODO commercialRegisterNumber */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* registrationInRegister */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.registrationInRegister'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'registrationInRegisterData'} {/* TODO registrationInRegister */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* registerNumber */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.registerNumber'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'registerNumberData'} {/* TODO registerNumber */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* abiCode */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.abiCode'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'abiCodeData'} {/* TODO abiCode */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* recipientCode */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.recipientCode'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'recipientCodeData'} {/* TODO recipientCode */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  <Typography
-                    variant="body1"
-                    sx={{
-                      fontWeight: 'fontWeightMedium',
-                      marginTop: 4,
-                      marginBottom: 2,
-                    }}
-                    ml={1}
-                  >
-                    {t(
-                      'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.protectionOfficerDetailData.title'
-                    )}
-                  </Typography>
-
-                  {/* protectionOfficerAddress */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.protectionOfficerDetailData.address'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'protectionOfficerAddressData'} {/* TODO protectionOfficerAddress */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* protectionOfficerMailPEC */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.protectionOfficerDetailData.mailPEC'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'protectionOfficerMailPEC'} {/* TODO protectionOfficerMailPECData */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* protectionOfficerMail */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.billingDataInfoSummarySection.billingDataInfoSummary.protectionOfficerDetailData.mail'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'protectionOfficerMailData'} {/* TODO protectionOfficerMail */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-                </Grid>
-              </Grid>
-            </Grid>
-          </Paper>
-          <Paper elevation={8} sx={{ borderRadius: theme.spacing(2) }}>
-            <Grid container sx={{ marginY: 4, marginX: 4 }}>
-              <Grid item xs={12}>
-                <TitleBox
-                  title={t(
-                    'requestToBeAnsweredPage.summaryStepSection.managerInfoSummarySection.title'
-                  )}
-                  variantTitle={'h6'}
-                  mtTitle={1}
-                  mbTitle={5}
-                />
-                <Divider />
-              </Grid>
-              <Grid item xs={12} mt={4}>
-                <Grid container spacing={1}>
-                  {/* name */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.managerInfoSummarySection.managerInfoSummary.name'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'nameData'} {/* TODO name */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* surname */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.managerInfoSummarySection.managerInfoSummary.surname'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'surnameData'} {/* TODO surname */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* taxCode */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.managerInfoSummarySection.managerInfoSummary.taxCode'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'taxCodeData'} {/* TODO taxCode */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* mailPEC */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.managerInfoSummarySection.managerInfoSummary.mailPEC'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'mailPECData'} {/* TODO mailPEC */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-                </Grid>
-              </Grid>
-            </Grid>
-          </Paper>
-          <Paper elevation={8} sx={{ borderRadius: theme.spacing(2) }}>
-            <Grid container sx={{ marginY: 4, marginX: 4 }}>
-              <Grid item xs={12}>
-                <TitleBox
-                  title={t(
-                    'requestToBeAnsweredPage.summaryStepSection.delegatesInfoSummarySection.title'
-                  )}
-                  variantTitle={'h6'}
-                  mtTitle={1}
-                  mbTitle={5}
-                />
-                <Divider />
-              </Grid>
-              <Grid item xs={12} mt={4}>
-                {/* TODO Add logic for show if present other delegates (using transcoded 
-                "requestToBeAnsweredPage.summaryStepSection.delegatesInfoSummarySection.delegatesInfoSummary.secondAdmin"
-                OR
-                "requestToBeAnsweredPage.summaryStepSection.delegatesInfoSummarySection.delegatesInfoSummary.thirdAdmin" */}
-                <Grid container spacing={1}>
-                  {/* name */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.delegatesInfoSummarySection.delegatesInfoSummary.name'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'nameData'} {/* TODO name */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* surname */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.delegatesInfoSummarySection.delegatesInfoSummary.surname'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'surnameData'} {/* TODO surname */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* taxCode */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.delegatesInfoSummarySection.delegatesInfoSummary.taxCode'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'taxCodeData'} {/* TODO taxCode */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-
-                  {/* mailPEC */}
-                  <Grid container item alignItems={'center'}>
-                    <Grid item xs={3}>
-                      <Typography sx={{ fontSize: 'fontSize' }}>
-                        {t(
-                          'requestToBeAnsweredPage.summaryStepSection.delegatesInfoSummarySection.delegatesInfoSummary.mailPEC'
-                        )}
-                      </Typography>
-                    </Grid>
-                    <Grid item xs={9} display="flex" alignItems={'center'}>
-                      <Typography sx={{ fontSize: 'fontSize', fontWeight: 'fontWeightMedium' }}>
-                        {'mailPECData'} {/* TODO mailPEC */}
-                      </Typography>
-                    </Grid>
-                  </Grid>
-                </Grid>
-              </Grid>
-            </Grid>
-          </Paper>
-        </Stack>
-        <Stack direction="row" justifyContent="space-between" alignItems="center" py={6} mb={6}>
-          <Stack>
-            <Button
-              variant="outlined"
-              color="error"
-              style={{ color: theme.palette.error.dark, borderColor: theme.palette.error.dark }}
-              onClick={() => setShowRejectPage(true)}
-            >
-              {/* TODO Add call for decline request */}
-              {t('requestToBeAnsweredPage.actions.declineButton')}
-            </Button>
-          </Stack>
-          {toValidateState && (
-            <Stack>
-              <Button
-                variant="contained"
-                sx={{ marginLeft: 3 }}
-                onClick={() => setShowConfirmPage(true)}
-              >
-                {/* TODO Add call for approve request */}
-                {t('requestToBeAnsweredPage.actions.approveButton')}
-              </Button>
-            </Stack>
-          )}
-        </Stack>
-      </Grid>
-    </Grid>
-  );
-}
